Bug 1091008 - Reimplement HasLowUndecodedData in terms of GetBuffered. r=cpearce
Since GetBuffered now has a sane implementation for MSE, this should
make this function sane for MSE as well.

UltraBlame original commit: 16f58d7e1e17254bb285368e14a140eb008c1a78

+++ dom/media/MediaDecoderStateMachine.h
@@ -427,7 +427,7 @@ protected:
   bool HasLowUndecodedData();
 
   
-  bool HasLowUndecodedData(double aUsecs);
+  bool HasLowUndecodedData(int64_t aUsecs);
